+++ bigdata/src/java/com/bigdata/isolation/UnisolatedBTree.java
@@ -495,7 +495,7 @@ public class UnisolatedBTree extends BTree implements IIsolatableIndex {
      * This method will include deleted entries in the key range in the returned
      * count.
      */
-    public int rangeCount(byte[] fromKey, byte[] toKey) {
+    public long rangeCount(byte[] fromKey, byte[] toKey) {
         
         return super.rangeCount(fromKey, toKey);
